support customize command

+++ src/main.rs
@@ -9,17 +9,26 @@ use std::{
 
 #[derive(Debug, Parser)]
 struct Options {
-    #[arg(short, long, default_value_t = 60)]
+    
+    /// to check activity interval
+    #[arg(short = 'i', default_value_t = 60)]
     check_interval: u64,
 
-    #[arg(short, long, default_value_t = 3600)]
+    /// how long to suspend after no user activity
+    #[arg(short = 's', default_value_t = 3600)]
     sec_to_suspend: u64,
 
-    #[arg(short, long, default_value_t = false)]
+    /// force to shuting down when suspend failed
+    #[arg(short = 'f', default_value_t = false)]
     force_shutdown: bool,
+
+    /// execute suspend action
+    #[arg(short = 'c', default_value = "/usr/bin/systemctl suspend")]
+    suspend_command: String,
 }
 
 const ASSUME_SUSPEND_FAILED_TOLERANCE: u64 = 3;
+const FORCE_SHUTDOWN_TIMEOUT: u64 = 60;
 
 fn main() {
     env_logger::init();
@@ -49,7 +58,7 @@ fn main() {
         if how_long_to_suspend <= 0 {
             warn!("suspend time was meet, system will be suspending");
             suspend_at = None;
-            match Command::new("/usr/bin/systemctl").arg("suspend").output() {
+            match must_to_command(&options.suspend_command).output() {
                 Err(e) => {
                     error!("execute suspend command error: {}", e);
                 }
@@ -71,9 +80,14 @@ fn main() {
                 // `systemctl suspend` may execute failed but no error returned
                 error!("instantaneous woken from suspend, assume suspend failed!");
                 if options.force_shutdown {
-                    thread::sleep(Duration::from_secs(3));
-                    warn!("force to shuting down machine");
-                    Command::new("/usr/bin/systemctl").arg("poweroff").exec();
+                    info!("system will be force shutdown at 60s");
+                    thread::sleep(Duration::from_secs(FORCE_SHUTDOWN_TIMEOUT));
+                    if is_any_user_logged_on() {
+                        warn!("detected user activity, stop to shutdown machine");
+                    } else {
+                        warn!("force to shuting down machine");
+                        Command::new("/usr/bin/systemctl").arg("poweroff").exec();
+                    }
                 }
             }
             continue;
@@ -83,6 +97,14 @@ fn main() {
         }
         thread::sleep(Duration::from_secs(how_long_to_suspend));
     }
+}
+
+fn must_to_command(str_cmd: &str) -> Command  {
+    let mut splits = str_cmd.trim().split_whitespace();
+    let cmd = splits.next().expect("command is missing");
+    let mut cmd = Command::new(cmd);
+    cmd.args(splits.collect::<Vec<&str>>());
+    cmd
 }
 
 fn is_any_user_logged_on() -> bool {
@@ -114,4 +136,23 @@ fn is_any_user_logged_on() -> bool {
         }
     }
     false
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::must_to_command;
+
+
+    #[test]
+    pub fn test_must_to_command() {
+        must_to_command("ls");
+        must_to_command("ls -alh");
+        must_to_command("/usr/bin/ls -l -a -h");
+    }
+
+    #[test]
+    #[should_panic]
+    pub fn test_panic_must_to_command() {
+        must_to_command("    ");
+    }
 }
